Pull out the series toggler thing
It should be shared between Series and Collection
